refactor: 使用者首頁控制器 目錄結構調整
