+++ resources/views/user/home.blade.php
